Add CDS 102 labs to course website
Labs
====

This update adds the following labs to the main CDS 101 course website. So far,
only the labs assigned during the Summer 2018 semester of CDS 102 are included,
which are,

* Reproducible research with RMarkdown and Github
* Exploration by visualization: the Galton dataset
* Wrangling sales data
* Tidying your dataset
* Mining the social web
* Speed of light
* Moneyball

Outdated methods in the labs have been updated

* Speed of light: Visualizations, p-values, and confidence intervals are now
  handled using functions from the infer package instead of being computed
  manually.
* Moneyball: The tidy() and glance() functions from the broom package no longer
  need to be manually converted to the tibble format. The manual function for
  plotting the reference line in Q-Q plots is no longer needed, as ggplot2 now
  has the function geom_qq_line.

The image files displayed in the lab instructions have also been added to the
repository.

Themes and templates
====================

Jinja2 templates were added for generating pages for the lab instructions. The
cards are sorted by page title in ascending order.

The Jinja2 template for the assignments page was updated so that the cards will
no longer turn gray. The assignments cards are also no longer sorted in reverse
order.

The Jinja2 template or the Materials page was updated so that the cards are
sorted by page title in ascending order.

The JavaScript that initializes the Wookmark card layout has been edited so that
different kinds of card layouts can be set by changing the class of the <ul>
HTML element. The layout for the labs cards was changed to have a maximum of two
cards per row.

An APS bibliography styling file and a references.bib Bibtex file was added to
the static folder. This is used to store bibliographic information that can be
referenced on any page of the website and instructions on how to style it upon
page generation. The citation style is the style used in the American Physical
Society journals. The references.bib file currently just stores the references
for the "Speed of Light" lab.

Environment
===========

* The Slack URL was updated to point to the Summer 2019 Slack workspace.
* Pelican was configured to use the template files for the lab instructions.
* A new button called "Labs" was added to the navigation bar in the Pelican
  configuration file.
* Pandoc arguments added to use the references.bib file for generating reference
  lists.