Update site formatting for cards and names for mini-homeworks
Webpage
=======

The names for each of the mini-homeworks now have the word "Mini-homework" in
their titles and are enumerated.

The lecture videos on the Materials page are now enclosed in div HTML elements
with the classes .video-container and .video-item.

The date metadata was fixed for the Module 6 exercises.

Themes and templates
====================

The Jinja2 template for naming the header of each assignment card on the
Assignments page was updated to use the label "Mini-Homework" instead of
"Mini-Assignment".

The Jinja2 template for generating the Materials cards page was updated so that
the Module cards are now the full width of the webpage, regardless of screen
size.

Created a Flexbox CSS container for lecture videos on the Materials page. Up to
three lecture videos will be displayed per row, depending on screen size.