Fix wording and formatting of titles above lecture videos
Webpage
=======

The titles above the lecture videos contained the name of the Module in them,
which was redundant, and so they have been removed.

In Module 7, the lecture video titles differ enough in length that some of the
titles are one-line but others are two lines. This makes the layout look
strange, so the height has been fixed for all the titles here to maintain
consistency.

Themes and templates
====================

The width per lecture video in the Flexbox container of the Materials cards was
expanded to 425px.