Publish the CDS 102 lab, Lab 00: Beginnings
Notes
=====

This is a new lab that students will complete in the first week of the course.
It will ensure that they have all their accounts registered and tools properly
configured for the duration of the semester. To check that students are actually
completing these steps, students will need to submit two items to Blackboard:

* A csv file that lists the versions of the packages that were installed to
  their RStudio Server account by following the Lab 00 instructions.
* A valid Twitter username that will be added to the team developer account for
  CDS 102.