Revise the instructions for mini-homeworks 2a and 2b
Notes
=====

The primary update to the instructions for mini-homework 2a and 2b edits the
text within the First-time RStudio Server configuration call-out box to instruct
students to follow the instructions within a section of Lab 00 to ensure their
RStudio Server environment is configured properly.

The instructions for mini-homework 2a were clarified so that students know they
should replace "FirstName LastName" with their own names and that it's okay that
the date on their knitted document will not match the date on the sample
document.

PDF and link icons were added to make it clearer where the clickable links are.

The sample HTML rendering of mini-homework 2b now uses the kbd HTML element for
representing keyboard keys instead of using monospace text.