fix(starters): don't commit *.Rproj files (#7)
The .gitignore files for all available starter templates have been
updated to prevent students from accidentally committing their
RStudio Project files to GitHub. When this was allowed, it led to
unnecessary merge conflicts when students work in groups, so it's
just easier to prevent it from happening in the first place.