content(starters): lab 11
Write new script to be distributed with lab 11 that automates the
configuration of rtweet. The script prompts users for their Twitter
API keys and access tokens and automatically takes care of the
rest. In addition, the number of exercises listed in the R Markdown
lab report template was expanded to seven.