content(labs): update and revise lab 11
An instructions preamble has been added to the beginning of Lab 8,
and the instructions for generating an authentication token for the
Twitter API were updated to be compatible with the new
developer.twitter.com dashboard. The lab instructions in general
have been edited and revised to improve clarity. One section of
instructions was converted into a new exercise. Other minor changes
include,

- Hint-like parentheticals were moved into callout boxes.
- The code for configuring the RStudio environment was moved to a
  script that automates the setup process.
- A list of relevant cheatsheets is linked at the end of the lab
  instructions.