content(labs): basketball and hot hands (#29)
The instructions and exercises for Lab 7 have been added to the
course website. An instructions preamble has been added to the
beginning of Lab 7, and the "About the dataset" section was
slightly revised to better match the way datasets are introduced in
the CDS 101 homeworks. The instructions were revised to use the
infer package to simulate flipping a coin and an independent
shooter and edited to improve clarify. Minor changes include,

- The "Additional questions" section has been renamed "Additional
  exercises"
- A list of relevant cheatsheets is linked at the end of the lab
  instructions